data-pipeline/requirements.txt
weaviate-client
pytube
yt-dlp
pafy
git+https://github.com/openai/whisper.git
torch
python-dotenv